backend/requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
google-generativeai>=0.7.2
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
Flask-Login==0.6.3
fpdf2==2.7.5
